updates to examples
